Refactor into scalp package
