+++ README.md
@@ -1,1 +1,9 @@
 # Scalp
+
+Run the bot with:
+
+```bash
+python -m scalp.main
+```
+
+The package contains modules for configuration, REST client and strategy helpers.
